docs: rm old TF provider, migrate links to Equinix
Mention the Terraform provider which is maintained instead of one which
has a flashing "unsupported" banner.

Use phrasing "Kubernetes CCM" instead of "Kubernetes driver for ... CCM"
because neither the Kubernetes official docs nor CCM docs refer to it
as a "driver".

Remove link to CSI, since it has been sunset.

Bypass the redirects in all remaining `packethost/*` links.

Signed-off-by: Jakub Bielecki <47531708+jabielecki@users.noreply.github.com>