add ClientOpt options
Add the ClientOpt, just classic functional options:
https://dave.cheney.net/2014/10/17/functional-options-for-friendly-apis

The immediate need is to provide means for callers to insert custom
headers into every request. For example, if callers need a third header
in addition to the typical X-Auth-Key and X-Consumer-Token pair, they
can now use WithHeader.

The existing possibility of customizing headers per every request
with DoRequestWithHeaders has the problem that for an existing code base
using packngo.DoRequest heavily, it becomes hard to add a header
to every request without a massive code modification.

I modified the constructor NewClient in a backward-compatible way to
use the functional options. The small difference in logic is that if the
builtin default base API URL turns out to be non-parseable:

  - now NewClient will return: nil Client, appropriate error
  - previously NewClient returned: nil Client, nil error

Also the order of reading env variables changed, which means that
from now on majority of the constructor code can benefit from
Client.debug (`$PACKNGO_DEBUG`). Previously it was known quite late
into the construction.

Signed-off-by: Jakub Bielecki <47531708+jabielecki@users.noreply.github.com>